host/client normalization for windrack & thunder
These moves will now work consistently regardless of who is regarded as host/client:

When Windrack is used point blank against someone who has a barrier, the wind gusts will also apply.

When Thunder/BDT is on the same panel as the character that it's targeting, it will always move in the direction that is considered "forward" by the character who spawned it.